make test program dump state to json
The test program now writes its state to json on stdout. This output is
intended to be captured in tests, to compare against pre-set environment
variable configurations.

Made otelinit.Config public so it can be tunneled through context and
picked up in the test program. There are other ways to accomplish this
but it might be useful to folks down the road for debugging. So now
the "otel-init-config" value in the context can be used to grab the
config without an increase in API surface other than returning context,
which like the last API change, it probably should have in the first place.
(I fully expect these changes to settle down by the time this PR merges.)

Signed-off-by: Amy Tobey <atobey@equinix.com>