add a full-loop test using otel-cli
Discovered I needed to add a wildcard match for especially trace and
span ids, but did it for the whole map while I was there. It's not the
prettiest approach but I tried writing a cmp.Comparator and it didn't
work out, so for now this gets the job done nicely.

Signed-off-by: Amy Tobey <atobey@equinix.com>